refactor(stripe): extract verifyStripeSignature into shared module (F-4)
Move the duplicated verifyStripeSignature function from both
payments/webhook/route.ts and billing/webhook/route.ts into a new
shared module at src/lib/stripe-signature.ts.

Both routes now import from the single source. No behavioral change —
same HMAC-SHA256 logic, same 5-minute timestamp tolerance, same
timingSafeEqual comparison.
