+++ src/app/api/billing/webhook/route.ts
@@ -1,8 +1,8 @@
 import { NextRequest } from "next/server";
 import { apiError, apiSuccess, apiInternalError } from "@/lib/api-response";
 import { getPlanByPriceId, type PlanSlug } from "@/lib/config/subscription-plans";
-import { hmacSha256Hex, timingSafeEqual } from "@/lib/crypto-utils";
 import { logger } from "@/lib/logger";
+import { verifyStripeSignature } from "@/lib/stripe-signature";
 import { createAdminClient } from "@/lib/supabase-server";
 import type { Json } from "@/lib/types/database";
 import { subscriptionWebhookEventSchema } from "@/lib/validations";
@@ -374,41 +374,5 @@ export async function POST(request: NextRequest) {
   } catch (err) {
     logger.warn("Failed to process billing webhook", { context: "billing/webhook", error: err });
     return apiInternalError("Failed to process webhook");
-  }
-}
-
-/**
- * Verify Stripe webhook signature using HMAC-SHA256.
- * Implements Stripe's signature verification without requiring the Stripe SDK.
- */
-async function verifyStripeSignature(
-  payload: string,
-  signatureHeader: string,
-  secret: string,
-): Promise<boolean> {
-  try {
-    const parts = signatureHeader.split(",");
-    let timestamp = "";
-    let signature = "";
-
-    for (const part of parts) {
-      const [key, value] = part.split("=");
-      if (key === "t") timestamp = value;
-      if (key === "v1") signature = value;
-    }
-
-    if (!timestamp || !signature) return false;
-
-    // Check timestamp tolerance (5 minutes)
-    const now = Math.floor(Date.now() / 1000);
-    if (Math.abs(now - parseInt(timestamp, 10)) > 300) return false;
-
-    const signedPayload = `${timestamp}.${payload}`;
-    const expectedSignature = await hmacSha256Hex(secret, signedPayload);
-
-    return timingSafeEqual(expectedSignature, signature);
-  } catch (err) {
-    logger.warn("Signature verification failed", { context: "billing/webhook", error: err });
-    return false;
   }
 }
