+++ src/lib/stripe-signature.ts
@@ -1,0 +1,43 @@
+import { hmacSha256Hex, timingSafeEqual } from "@/lib/crypto-utils";
+import { logger } from "@/lib/logger";
+
+/**
+ * Verify Stripe webhook signature using HMAC-SHA256.
+ * Implements Stripe's signature verification without requiring the Stripe SDK.
+ *
+ * Shared by payments/webhook and billing/webhook routes.
+ */
+export async function verifyStripeSignature(
+  payload: string,
+  signatureHeader: string,
+  secret: string,
+): Promise<boolean> {
+  try {
+    const parts = signatureHeader.split(",");
+    let timestamp = "";
+    let signature = "";
+
+    for (const part of parts) {
+      const [key, value] = part.split("=");
+      if (key === "t") timestamp = value;
+      if (key === "v1") signature = value;
+    }
+
+    if (!timestamp || !signature) return false;
+
+    // Check timestamp tolerance (5 minutes)
+    const now = Math.floor(Date.now() / 1000);
+    if (Math.abs(now - parseInt(timestamp, 10)) > 300) return false;
+
+    const signedPayload = `${timestamp}.${payload}`;
+    const expectedSignature = await hmacSha256Hex(secret, signedPayload);
+
+    return timingSafeEqual(expectedSignature, signature);
+  } catch (err) {
+    logger.warn("Stripe signature verification failed", {
+      context: "stripe-signature",
+      error: err,
+    });
+    return false;
+  }
+}
